move external keys for broadcasts
Before the external keys for broadcasts were managed in the polymorphic association `external_keys`. Since a Broadcast is a single artifact within a station anyway we can omit this kind of duplication of data.

Table `broadcasts` receives a column `external_key`, data is being moved from `external_keys` to `broadcasts`.